refactor(workbench): lift Workflow shell into the design system
User direction: "each app will have a workflow defined" — so the
shell needs to be reusable, not buried in a campus-specific file.

New in `@klorad/design-system`:
  - `WorkflowPanel` — the full shell: optional title, tab bar,
    description band, scrollable content. Accepts a `steps: WorkflowStep[]`
    array; the active step's `render()` runs in the content area.
  - `WorkflowTabBar` — horizontal pills (icon-above-label) with the
    accent-soft / accent-on-hover treatment we landed on.
  - `WorkflowListItem` — the "empty interior + 1px grey border at
    rest, border-accent on hover, accent-soft fill on select" card
    container. Use when a row has internal action buttons.
  - `WorkflowListButton` — same styling but the root *is* a button.
    Use when the whole row is one click target.
  - `WorkflowDrillHeader` — "← back to {parent}" affordance.
  - `useWorkflowDrill<T>()` — drill-stack hook. Returns
    `{ path, depth, push, pop, reset }`. Type-parameterised for
    verticals that need richer level state than a string id.

Campus now wires these in `apps/campus/lib/workbench/views/workflow/`:
  - `index.tsx` (60 LOC)  — declares the three steps + mounts
    `<WorkflowPanel>`. The shape every future vertical copies.
  - `location-step.tsx`   — Campus-specific (Mapbox center capture,
    `api.setLocation` call).
  - `buildings-step.tsx`  — Building → Floor → Room drill via
    `useWorkflowDrill`. Uses `WorkflowListItem` + `WorkflowDrillHeader`.
  - `pois-step.tsx`       — Live-filtered list using
    `WorkflowListButton`.
  - `shared.tsx`          — Campus-internal helpers reused across
    steps (EmptyHint, SearchPill).

Net: ~900-LOC monolith split into one mount file + three focused
step files, with the reusable bits (≈250 LOC) lifted into the DS.

Defining the next vertical's workflow is now a 5-line setup:

```ts
<WorkflowPanel
  title="Workflow"
  steps={[
    { id: "routes",   label: "Routes",   icon: RouteIcon,   description: "…", render: () => <RoutesStep /> },
    { id: "vehicles", label: "Vehicles", icon: VehicleIcon, description: "…", render: () => <VehiclesStep /> },
    { id: "stops",    label: "Stops",    icon: StopIcon,    description: "…", render: () => <StopsStep /> },
  ]}
/>
```

…plus three step component files. The shell, tabs, drill nav,
list-item styling, hover / select states — all free.
